+++ src/reduxs/application/actions.ts
@@ -1,14 +1,21 @@
 import { createAction, ActionFunctionAny, Action } from 'redux-actions';
+import {WindowOptions} from '@albertli90/redux-openfin';
+
 import { IApplicationNewSnackbarOption, IApplicationSetSnackbarStatusOption, IApplicationCloseSnackBarOption } from './types';
 
+export const APPLICATION_SET_LOADING_MSG:string = 'APPLICATION_SET_LOADING_MSG';
 export const APPLICATION_STARTED:string = 'APPLICATION_STARTED';
 export const APPLICATION_CHILD_STARTED:string = 'APPLICATION_CHILD_STARTED';
+export const APPLICATION_NOTIFICATION_STARTED:string = 'APPLICATION_NOTIFICATION_STARTED';
 export const APPLICATION_READY:string = 'APPLICATION_READY';
 export const APPLICATION_DRAWER_TOGGLE:string = 'APPLICATION_DRAWER_TOGGLE';
 export const APPLICATION_TOGGLE_WINDOW_STATE:string = 'APPLICATION_TOGGLE_WINDOW_STATE';
 
+export const applicationSetLoadingMsg:ActionFunctionAny<Action<{loadingMsg:string}>> =
+    createAction<{loadingMsg:string},string>(APPLICATION_SET_LOADING_MSG,(loadingMsg:string)=>({loadingMsg}));
 export const applicationStarted:ActionFunctionAny<Action<void>> = createAction(APPLICATION_STARTED);
 export const applicationChildStarted:ActionFunctionAny<Action<void>> = createAction(APPLICATION_CHILD_STARTED);
+export const applicationNotificationStarted:ActionFunctionAny<Action<void>> = createAction(APPLICATION_NOTIFICATION_STARTED);
 export const applicationReady:ActionFunctionAny<Action<void>> = createAction(APPLICATION_READY);
 export const applicationDrawerToggle:ActionFunctionAny<Action<void>> = createAction(APPLICATION_DRAWER_TOGGLE);
 export const applicationToogleWindowState:ActionFunctionAny<Action<void>> = createAction(APPLICATION_TOGGLE_WINDOW_STATE);
@@ -40,10 +47,12 @@ export const applicationUpdateDockStatus:ActionFunctionAny<Action<{docked:boolea
 export const APPLICATION_LAUNCH_BAR_TOGGLE = 'APPLICATION_LAUNCH_BAR_TOGGLE';
 export const APPLICATION_LAUNCH_BAR_TOGGLE_COLLAPSE = 'APPLICATION_LAUNCH_BAR_TOGGLE_COLLAPSE';
 export const APPLICATION_LAUNCH_NEW_WINDOW = 'APPLICATION_LAUNCH_NEW_WINDOW';
+export const APPLICATION_LAUNCH_BAR_CLOSE = 'APPLICATION_LAUNCH_BAR_CLOSE';
 export const applicationLaunchBarToggle:ActionFunctionAny<Action<void>> = createAction(APPLICATION_LAUNCH_BAR_TOGGLE);
 export const applicationLaunchBarToggleCollapse:ActionFunctionAny<Action<void>> = createAction(APPLICATION_LAUNCH_BAR_TOGGLE_COLLAPSE);
-export const applicationLaunchNewWindow:ActionFunctionAny<Action<any>>
-    = createAction(APPLICATION_LAUNCH_NEW_WINDOW,(appJson:any)=>(appJson));
+export const applicationLaunchNewWindow:ActionFunctionAny<Action<Partial<WindowOptions>>>
+    = createAction(APPLICATION_LAUNCH_NEW_WINDOW,(appJson:Partial<WindowOptions>)=>(appJson));
+export const applicationLaunchBarClose:ActionFunctionAny<Action<void>> = createAction(APPLICATION_LAUNCH_BAR_CLOSE);
 
 // network
 export const APPLICATION_NETWORK_ONLINE = 'APPLICATION_NETWORK_ONLINE';