+++ src/routes/index.ts
@@ -2,6 +2,7 @@ import { RouteItem } from './Base';
 import Loading from '../layouts/Loading/Loading';
 import Dashboard from '../layouts/Dashboard/Dashboard';
 import ChildWindow from '../layouts/ChildWindow/ChildWindow';
+import Notification from '../layouts/Notification/Notification';
 import LaunchBar from '../layouts/LaunchBar/LaunchBar';
 
 export * from './Base';
@@ -20,6 +21,10 @@ const indexRoutes:RouteItem[] = [
     {
         path:'/childWindow',
         component: ChildWindow,
+    },
+    {
+        path:'/notification',
+        component: Notification,
     },
     {
         path:'/launchBar',