+++ src/utils/configureStore.ts
@@ -1,7 +1,6 @@
 import { applyMiddleware, createStore, compose } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 import { createOpenfinMiddleware } from '@albertli90/redux-openfin';
-import { ChannelType } from '@albertli90/redux-openfin/init';
 
 import rootReducer, {IRootState} from '../reduxs';
 import rootSaga from '../reduxs/sagas';